Conflicts:
	whois-api/src/main/java/net/ripe/db/whois/api/wsearch/LogSearchService.java
	whois-client/src/main/java/net/ripe/db/whois/common/rpsl/RpslAttribute.java
	whois-client/src/test/java/net/ripe/db/whois/common/rpsl/RpslObjectTest.java
	whois-commons/src/main/java/net/ripe/db/whois/common/rpsl/RpslObjectBuilder.java
	whois-commons/src/main/resources/pending_schema.sql
	whois-commons/src/main/resources/scheduler_schema.sql
	whois-commons/src/main/resources/whois_schema.sql
	whois-commons/src/test/java/net/ripe/db/whois/common/dao/jdbc/DatabaseHelper.java
	whois-commons/src/test/java/net/ripe/db/whois/common/support/DummyWhoisClient.java
	whois-endtoend/src/test/java/net/ripe/db/whois/WhoisFixture.java
	whois-query/src/test/java/net/ripe/db/whois/query/integration/SimpleTestIntegration.java
	whois-update/src/main/java/net/ripe/db/whois/update/keycert/PgpPublicKeyWrapper.java
	whois-update/src/main/java/net/ripe/db/whois/update/keycert/X509CertificateWrapper.java

+++ whois-api/src/main/java/net/ripe/db/whois/api/freetext/FreeTextIndex.java
@@ -1,7 +1,7 @@
 package net.ripe.db.whois.api.freetext;
 
+import com.google.common.base.CharMatcher;
 import com.google.common.base.Stopwatch;
-import com.google.common.collect.Lists;
 import com.google.common.collect.Maps;
 import com.google.common.collect.Sets;
 import net.ripe.db.whois.api.search.IndexTemplate;
@@ -12,19 +12,19 @@ import net.ripe.db.whois.common.domain.serials.SerialEntry;
 import net.ripe.db.whois.common.rpsl.AttributeType;
 import net.ripe.db.whois.common.rpsl.RpslAttribute;
 import net.ripe.db.whois.common.rpsl.RpslObject;
+import org.apache.commons.lang.StringUtils;
 import org.apache.lucene.analysis.Analyzer;
 import org.apache.lucene.document.Document;
 import org.apache.lucene.document.Field;
 import org.apache.lucene.document.FieldType;
-import org.apache.lucene.facet.index.FacetFields;
-import org.apache.lucene.facet.taxonomy.CategoryPath;
+import org.apache.lucene.facet.FacetField;
+import org.apache.lucene.facet.FacetsConfig;
 import org.apache.lucene.facet.taxonomy.TaxonomyWriter;
 import org.apache.lucene.index.IndexWriter;
 import org.apache.lucene.index.IndexWriterConfig;
 import org.apache.lucene.index.Term;
 import org.apache.lucene.util.Version;
 import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.beans.factory.annotation.Value;
@@ -43,10 +43,18 @@ import java.sql.SQLException;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
+import java.util.concurrent.ArrayBlockingQueue;
+import java.util.concurrent.ExecutorService;
+import java.util.concurrent.ThreadPoolExecutor;
+import java.util.concurrent.TimeUnit;
 
+import static com.google.common.collect.Lists.newArrayListWithExpectedSize;
+import static org.slf4j.LoggerFactory.getLogger;
+
+// TODO: instead of relying on scattered 'if (StringUtils.isBlank(indexDir) {...}', we should use profiles/...
 @Component
 public class FreeTextIndex extends RebuildableIndex {
-    private static final Logger LOGGER = LoggerFactory.getLogger(FreeTextIndex.class);
+    private static final Logger LOGGER = getLogger(FreeTextIndex.class);
 
     private static final int INDEX_UPDATE_INTERVAL_IN_SECONDS = 60;
 
@@ -54,8 +62,8 @@ public class FreeTextIndex extends RebuildableIndex {
     static final String PRIMARY_KEY_FIELD_NAME = "primary-key";
     static final String LOOKUP_KEY_FIELD_NAME = "lookup-key";
 
-    static final Analyzer QUERY_ANALYZER = new FreeTextAnalyzer(Version.LUCENE_41, FreeTextAnalyzer.Operation.QUERY);
-    static final Analyzer INDEX_ANALYZER = new FreeTextAnalyzer(Version.LUCENE_41, FreeTextAnalyzer.Operation.INDEX);
+    static final Analyzer QUERY_ANALYZER = new FreeTextAnalyzer(FreeTextAnalyzer.Operation.QUERY);
+    static final Analyzer INDEX_ANALYZER = new FreeTextAnalyzer(FreeTextAnalyzer.Operation.INDEX);
 
     static final String[] FIELD_NAMES;
 
@@ -65,7 +73,7 @@ public class FreeTextIndex extends RebuildableIndex {
     private static final FieldType INDEXED_NOT_TOKENIZED;
 
     static {
-        final List<String> names = Lists.newArrayListWithExpectedSize(AttributeType.values().length);
+        final List<String> names = newArrayListWithExpectedSize(AttributeType.values().length);
         for (final AttributeType attributeType : AttributeType.values()) {
             if (!SKIPPED_ATTRIBUTES.contains(attributeType)) {
                 names.add(attributeType.getName());
@@ -87,28 +95,28 @@ public class FreeTextIndex extends RebuildableIndex {
         INDEXED_NOT_TOKENIZED.freeze();
     }
 
-    public static final int MAX_UPDATE_BACKLOG = 25000;
-
     private final JdbcTemplate jdbcTemplate;
     private final String source;
+    private final FacetsConfig facetsConfig;
 
     @Autowired
     FreeTextIndex(
             @Qualifier("whoisSlaveDataSource") final DataSource dataSource,
             @Value("${whois.source}") final String source,
-            @Value("${dir.freetext.index}") final String indexDir) {
+            @Value("${dir.freetext.index:}") final String indexDir) {
 
         super(LOGGER, indexDir);
 
         this.jdbcTemplate = new JdbcTemplate(dataSource);
         this.source = source;
+        this.facetsConfig = new FacetsConfig();
     }
 
     @PostConstruct
     public void init() {
-        super.init(new IndexWriterConfig(Version.LUCENE_41, INDEX_ANALYZER)
-                .setOpenMode(IndexWriterConfig.OpenMode.CREATE_OR_APPEND),
-
+        if (StringUtils.isBlank(indexDir)) return;
+        super.init(new IndexWriterConfig(Version.LUCENE_4_10_0, INDEX_ANALYZER)
+                        .setOpenMode(IndexWriterConfig.OpenMode.CREATE_OR_APPEND),
                 new IndexTemplate.WriteCallback() {
                     @Override
                     public void write(final IndexWriter indexWriter, final TaxonomyWriter taxonomyWriter) throws IOException {
@@ -117,30 +125,22 @@ public class FreeTextIndex extends RebuildableIndex {
                         } else {
                             final Map<String, String> commitData = indexWriter.getCommitData();
                             final String committedSource = commitData.get("source");
+
                             if (!source.equals(committedSource)) {
                                 LOGGER.warn("Index {} has invalid source: {}, rebuild", indexDir, committedSource);
                                 rebuild(indexWriter, taxonomyWriter);
+                                return;
                             }
 
                             if (!commitData.containsKey("serial")) {
                                 LOGGER.warn("Index {} is missing serial, rebuild", indexDir);
                                 rebuild(indexWriter, taxonomyWriter);
-                            }
-
-                            final int indexSerial = Integer.parseInt(commitData.get("serial"));
-                            final int serial = JdbcRpslObjectOperations.getSerials(jdbcTemplate).getEnd();
-                            if (serial < indexSerial) {
-                                LOGGER.warn("Index serial ({}) higher than database serial ({}), rebuild", indexSerial, serial);
-                                rebuild(indexWriter, taxonomyWriter);
-                            }
-
-                            if (serial - MAX_UPDATE_BACKLOG > indexSerial) {
-                                LOGGER.warn("Index serial ({}) too far behind database serial ({}), rebuild", indexSerial, serial);
-                                rebuild(indexWriter, taxonomyWriter);
+                                return;
                             }
                         }
                     }
-                });
+                }
+        );
     }
 
     @PreDestroy
@@ -149,14 +149,21 @@ public class FreeTextIndex extends RebuildableIndex {
     }
 
     protected void rebuild(final IndexWriter indexWriter, final TaxonomyWriter taxonomyWriter) throws IOException {
+        if (StringUtils.isBlank(indexDir)) return;
+
         indexWriter.deleteAll();
         final int maxSerial = JdbcRpslObjectOperations.getSerials(jdbcTemplate).getEnd();
 
+        // sadly Executors don't offer a bounded/blocking submit() implementation
+        int numThreads = Runtime.getRuntime().availableProcessors();
+        final ArrayBlockingQueue<Runnable> workQueue = new ArrayBlockingQueue<>(numThreads * 64);
+        final ExecutorService executorService = new ThreadPoolExecutor(numThreads, numThreads,
+                0L, TimeUnit.MILLISECONDS, workQueue, new ThreadPoolExecutor.CallerRunsPolicy());
+
         JdbcStreamingHelper.executeStreaming(jdbcTemplate, "" +
-                "SELECT object_id, object " +
-                "FROM last " +
-                "WHERE sequence_id != 0 " +
-                "AND object_type != 100 ",
+                        "SELECT object_id, object " +
+                        "FROM last " +
+                        "WHERE sequence_id != 0 ",
                 new ResultSetExtractor<Void>() {
                     private static final int LOG_EVERY = 500000;
 
@@ -165,38 +172,32 @@ public class FreeTextIndex extends RebuildableIndex {
                         int nrIndexed = 0;
 
                         while (rs.next()) {
-                            final int objectId = rs.getInt(1);
+                            executorService.submit(new DatabaseObjectProcessor(rs.getInt(1), rs.getBytes(2), indexWriter, taxonomyWriter));
 
-                            final RpslObject object;
-                            try {
-                                object = RpslObject.parse(objectId, rs.getBytes(2));
-                            } catch (RuntimeException e) {
-                                LOGGER.warn("Unable to parse object with id: {}", objectId, e);
-                                continue;
-                            }
-
-                            try {
-                                addEntry(indexWriter, taxonomyWriter, object);
-
-                                if (++nrIndexed % LOG_EVERY == 0) {
-                                    LOGGER.info("Indexed {} objects", nrIndexed);
-                                }
-                            } catch (IOException e) {
-                                throw new IllegalStateException("Indexing", e);
+                            if (++nrIndexed % LOG_EVERY == 0) {
+                                LOGGER.info("Indexed {} objects", nrIndexed);
                             }
                         }
 
                         LOGGER.info("Indexed {} objects", nrIndexed);
                         return null;
                     }
-                });
+                }
+        );
 
+        executorService.shutdown();
+        try {
+            executorService.awaitTermination(1, TimeUnit.DAYS);
+        } catch (InterruptedException e) {
+            LOGGER.error("shutdown", e);
+        }
 
         updateMetadata(indexWriter, source, maxSerial);
     }
 
     @Scheduled(fixedDelay = INDEX_UPDATE_INTERVAL_IN_SECONDS * 1000)
     public void scheduledUpdate() {
+        if (StringUtils.isBlank(indexDir)) return;
         update();
     }
 
@@ -209,13 +210,13 @@ public class FreeTextIndex extends RebuildableIndex {
             LOGGER.warn("Index serial ({}) higher than database serial ({}), rebuilding", last, end);
             rebuild(indexWriter, taxonomyWriter);
         } else if (last < end) {
-            LOGGER.debug("Updating index {} to {}", indexDir, end);
+            LOGGER.debug("Updating index {} from {} to {}", indexDir, last, end);
 
-            final Stopwatch stopwatch = new Stopwatch().start();
+            final Stopwatch stopwatch = Stopwatch.createStarted();
             for (int serial = last + 1; serial <= end; serial++) {
-                final SerialEntry serialEntry = JdbcRpslObjectOperations.getById(jdbcTemplate, serial);
+                final SerialEntry serialEntry = JdbcRpslObjectOperations.getSerialEntry(jdbcTemplate, serial);
                 if (serialEntry == null) {
-                    // TODO: [AH] suboptimal; there could be big gaps in serial entries. we should have a getNextId() call instead, SELECT()ing on serial_id > serial
+                    // suboptimal;there could be big gaps in serial entries.
                     continue;
                 }
                 final RpslObject rpslObject = serialEntry.getRpslObject();
@@ -252,18 +253,51 @@ public class FreeTextIndex extends RebuildableIndex {
 
         for (final RpslAttribute attribute : rpslObject.getAttributes()) {
             if (!SKIPPED_ATTRIBUTES.contains(attribute.getType())) {
-                document.add(new Field(attribute.getKey(), attribute.getValue().trim(), INDEXED_AND_TOKENIZED));
+                document.add(new Field(attribute.getKey(), sanitise(attribute.getValue().trim()), INDEXED_AND_TOKENIZED));
             }
         }
 
-        final CategoryPath categoryPath = new CategoryPath(OBJECT_TYPE_FIELD_NAME, rpslObject.getType().getName());
-        final FacetFields facetFields = new FacetFields(taxonomyWriter);
-        facetFields.addFields(document, Lists.newArrayList(categoryPath));
+        document.add(new FacetField(OBJECT_TYPE_FIELD_NAME, rpslObject.getType().getName()));
 
-        indexWriter.addDocument(document);
+        indexWriter.addDocument(facetsConfig.build(taxonomyWriter, document));
+    }
+
+    private static String sanitise(final String value) {
+        return CharMatcher.JAVA_ISO_CONTROL.removeFrom(value);
     }
 
     private void deleteEntry(final IndexWriter indexWriter, final RpslObject rpslObject) throws IOException {
         indexWriter.deleteDocuments(new Term(PRIMARY_KEY_FIELD_NAME, Integer.toString(rpslObject.getObjectId())));
+    }
+
+    final class DatabaseObjectProcessor implements Runnable {
+        final int objectId;
+        final byte[] object;
+        final IndexWriter indexWriter;
+        final TaxonomyWriter taxonomyWriter;
+
+        private DatabaseObjectProcessor(int objectId, byte[] object, IndexWriter indexWriter, TaxonomyWriter taxanomyWriter) {
+            this.objectId = objectId;
+            this.object = object;
+            this.indexWriter = indexWriter;
+            this.taxonomyWriter = taxanomyWriter;
+        }
+
+        @Override
+        public void run() {
+            final RpslObject rpslObject;
+            try {
+                rpslObject = RpslObject.parse(objectId, object);
+            } catch (RuntimeException e) {
+                LOGGER.warn("Unable to parse object with id: {}", objectId, e);
+                return;
+            }
+
+            try {
+                addEntry(indexWriter, taxonomyWriter, rpslObject);
+            } catch (IOException e) {
+                throw new IllegalStateException("Indexing", e);
+            }
+        }
     }
 }
